# docs/collection/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sistema de Autoatendimento da Lanchonete",
        "description": "Documentacao automatica via Swagger e Redoc",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pagamento/": {
            "get": {
                "tags": ["pagamento"],
                "summary": "Listar todos os pagamentos realizado",
                "operationId": "listar_pagamentos_pagamento__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PagamentoResponseSchema"
                                    },
                                    "type": "array",
                                    "title": "Response Listar Pagamentos Pagamento  Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao buscar os pagamentos"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["pagamento"],
                "summary": "Criar pagamento do pedido",
                "operationId": "efetuar_pagamento_pedido_pagamento__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PagamentoCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao salvar o pagamento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pagamento/{codigo_pagamento}": {
            "get": {
                "tags": ["pagamento"],
                "summary": "Buscar Pagamento",
                "operationId": "buscar_pagamento_pagamento__codigo_pagamento__get",
                "parameters": [{
                    "name": "codigo_pagamento",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "title": "Codigo Pagamento"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao buscar o pagamento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pagamento não encontrado"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["pagamento"],
                "summary": "Atualizar Pagamento",
                "operationId": "atualizar_pagamento_pagamento__codigo_pagamento__put",
                "parameters": [{
                    "name": "codigo_pagamento",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "title": "Codigo Pagamento"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PagamentoAtualizaSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pagamento não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao atualizar o pagamento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["pagamento"],
                "summary": "Deletar Pagamento",
                "operationId": "deletar_pagamento_pagamento__codigo_pagamento__delete",
                "parameters": [{
                    "name": "codigo_pagamento",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "title": "Codigo Pagamento"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Pagamento deletado com sucesso",
                        "content": {
                            "application/json": {
                                "example": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Pagamento não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao remover o pagamento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "PagamentoAtualizaSchema": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": ["status"],
                "title": "PagamentoAtualizaSchema"
            },
...
            "PagamentoCreateSchema": {
                "properties": {
                    "pedido_id": {
                        "type": "integer",
                        "title": "Pedido Id"
                    }
                },
                "type": "object",
                "required": ["pedido_id"],
                "title": "PagamentoCreateSchema"
            },
            "PagamentoResponseSchema": {
                "properties": {
                    "pedido_id": {
                        "type": "integer",
                        "title": "Pedido Id"
                    },
                    "codigo_pagamento": {
                        "type": "string",
                        "title": "Codigo Pagamento"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": ["pedido_id", "codigo_pagamento", "status"],
                "title": "PagamentoResponseSchema"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [{
                                "type": "string"
                            }, {
                                "type": "integer"
                            }]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}